requirements.txt
# FastAPI-Flow 프로젝트를 위한 확장 의존성

# 기본 웹 프레임워크
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
python-multipart==0.0.9

# 보안 관련 라이브러리
python-jose[cryptography]==3.3.0  # JWT 토큰 처리
passlib[bcrypt]==1.7.4             # 비밀번호 해싱
pydantic-settings==2.2.1          # 설정 관리

# 세션 관리를 위한 Redis (선택적)
aioredis==2.0.1                   # Redis 비동기 클라이언트

# 파일 처리
aiofiles==23.2.1                  # 비동기 파일 처리

# 로깅 강화
structlog==24.1.0                 # 구조적 로깅
python-json-logger==2.0.7         # JSON 포맷 로깅

# 데이터베이스 (선택적)
# SQLAlchemy==2.0.29              # ORM
# alembic==1.13.1                 # 데이터베이스 마이그레이션

# 템플릿 (선택적)
# jinja2==3.1.3                   # HTML 템플릿 

# 프로덕션 서버 (선택적)
# gunicorn==21.2.0                # 프로덕션 WSGI 서버